# analyses/OpenAPI_ColisApp.yaml
openapi: 3.0.0
info:
  title: ColisApp API
  description: Documentation de l'API RESTful pour l'application de gestion des envois ColisApp.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local
  - url: https://colisapp.example.com/api/v1
    description: Serveur de production
security:
  - BearerAuth: []
...
paths:
...
  /(auth)/reset-password:
    post:
      summary: Réinitialiser le mot de passe de l'utilisateur
      description: Réinitialise le mot de passe de l'utilisateur à l'aide d'un token de réinitialisation valide.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Mot de passe mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Token ou mot de passe invalide
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
  /addresses:
    get:
      summary: Récupérer toutes les adresses
      description: Récupère une liste de toutes les adresses dans le système.
      tags:
        - Addresses
      responses:
        '200':
          description: Liste d'adresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Créer une nouvelle adresse
      description: Crée un nouvel enregistrement d'adresse dans le système.
      tags:
        - Addresses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAddressDto'
      responses:
        '201':
          description: Adresse créée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  address:
                    $ref: '#/components/schemas/AddressDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /addresses/{id}:
    get:
      summary: Récupérer une adresse par ID
      description: Récupère un enregistrement d'adresse unique par son ID.
      tags:
        - Addresses
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'adresse à récupérer
      responses:
        '200':
          description: Détails de l'adresse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Mettre à jour une adresse par ID
      description: Met à jour un enregistrement d'adresse existant par son ID.
      tags:
        - Addresses
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'adresse à mettre à jour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressDto'
      responses:
        '200':
          description: Adresse mise à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  address:
                    $ref: '#/components/schemas/AddressDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Supprimer une adresse par ID
      description: Supprime un enregistrement d'adresse par son ID.
      tags:
        - Addresses
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID de l'adresse à supprimer
      responses:
        '200':
          description: Adresse supprimée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  address:
                    $ref: '#/components/schemas/AddressDto'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/dashboard:
    get:
      summary: Récupérer les données du tableau de bord Admin
      description: Récupère les données du tableau de bord en fonction du rôle de l'administrateur authentifié (SUPER_ADMIN, AGENCY_ADMIN, ACCOUNTANT).
      tags:
        - Admin
        - Dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Données du tableau de bord
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummaryDto'
        '400':
          description: Rôle non reconnu
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Accès non autorisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
  /agencies/admin-agencies:
    get:
      summary: Récupérer les agences pour l'administrateur
      description: Récupère une liste d'agences en fonction du rôle de l'administrateur authentifié. SUPER_ADMIN voit toutes les agences, AGENCY_ADMIN voit les agences qui lui sont associées.
      tags:
        - Agencies
        - Admin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Numéro de page
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Nombre d'éléments par page
        - in: query
          name: search
          schema:
            type: string
          description: Terme de recherche
        - in: query
          name: sortKey
          schema:
            type: string
            default: name
          description: Clé de tri
        - in: query
          name: sortDir
          schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          description: Direction de tri
      responses:
        '200':
          description: Liste d'agences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgencyDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /agencies/create-agency:
    post:
      summary: Créer une nouvelle agence
      description: Crée un nouvel enregistrement d'agence. Accessible uniquement par SUPER_ADMIN.
      tags:
        - Agencies
        - Admin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgencyDto'
      responses:
        '200':
          description: Agence créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgencyDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Non autorisé
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Not found
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid data provided
...
    InternalServerError:
      description: Erreur interne du serveur
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
    Forbidden:
      description: Accès refusé
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Forbidden
  schemas:
...
    AgencyDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        vatNumber:
          type: string
        addressId:
          type: integer
        address:
          $ref: '#/components/schemas/AddressDto'
        capacity:
          type: integer
          nullable: true
        availableSlots:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - addressId
        - createdAt
        - updatedAt
    CreateAgencyDto:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        vatNumber:
          type: string
        addressId:
          type: integer
        address:
          $ref: '#/components/schemas/AddressDto'
        capacity:
          type: integer
          nullable: true
        availableSlots:
          type: integer
          nullable: true
      required:
        - name
        - addressId
...
    AddressDto:
      type: object
      properties:
        id:
          type: integer
        street:
          type: string
        complement:
          type: string
          nullable: true
        streetNumber:
          type: string
          nullable: true
        boxNumber:
          type: string
          nullable: true
        cityId:
          type: integer
        city:
          $ref: '#/components/schemas/CityDto'
      required:
        - street
        - city
...
    CityDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          $ref: '#/components/schemas/CountryDto'
      required:
        - id
        - name
        - country

    # Country DTOs
    CountryDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        iso2:
          type: string
          nullable: true
        iso3:
          type: string
          nullable: true
        phonecode:
          type: string
          nullable: true
        capital:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
      required:
        - id
        - name